Create README.md
The repository contains the DANSYS spreadsheet, a data analysis spreadsheet for LibreOffice Calc, It is a growing project with the basic DANSYS spreadsheet finished but the extended version, DANSYSX, to be updated. There are also  user guides, also being updated over the long haul. Interactive educational texts called LabBooks are also being created on DANSYS documents. The repository also contains video and audio materials for my website, The Therian Timeline, and blog, Adventuring: The Bear Creek Commentaries, a website for lifelong learners.